framework: Add mustpass option for the Vulkan CTS test runner
This option was added mostly to being able to run the mustpass
case list included on Vulkan CTS repository. As that file is in
txt format, we assume that the case list will be in txt format.

That is not the case for the other deqp-xxx profiles, that uses
a xml format. In any case, it is properly documented on the
piglit.conf.example

v2: allow to receive a txt file with just the name of the test,
    without any formatting