gl-4.5/named-framebuffer-draw-buffers-errors: add a new test
Equivalent to the already existing gl-3.1/draw-buffers-errors and
spec/gles-2.0/draw-buffers, but using the 4.5 API. The errors are
the same that with DrawBuffers so we are testing this method too,
although indirectly.

Used as reference OpenGL 4.5 spec, section 17.4.1 "Selecting Buffers
for Writing", page 490 (511 on the PDF).

Note that the behaviour is slightly different on compared with 4.4 and
lower, but we are testing only 4.5, as is the most recent 4.x spec.